ai-engine/requirements.txt
# --- 1. Deep Learning (PyTorch + CUDA 12.1) ---
# 윈도우에서 RTX 4060용 GPU 버전을 받기 위한 필수 링크
--find-links https://download.pytorch.org/whl/cu121/torch_stable.html
torch
torchvision
torchaudio

# --- 2. Hugging Face (AI 모델) ---
# 최신 버전의 PyTorch 보안 강제 오류(ValueError)를 피하기 위해 버전 고정
transformers==4.45.2
datasets
accelerate
tokenizers

# 토크나이저 필수 의존성 (누락 시 에러 발생)
sentencepiece
protobuf

# --- 3. Image Processing (데이터 생성/증강) ---
opencv-python
Pillow
albumentations

# --- 4. Data & Utilities ---
numpy
pandas
openpyxl
Faker
tqdm
scikit-learn
matplotlib
fuzzywuzzy
requests
python-dotenv